--- app/agents/helper.py
@@ -1,3 +1,0 @@
-def load_instructions(file_path: str) -> str:
-    with open(file_path, "r") as file:
-        return file.read()
